refactor: fix parsing directory structure
