+++ src/parsing/CMakeLists.txt
@@ -1,17 +1,12 @@
 target_sources(dndmanager
     PRIVATE
-        content_parser.cpp
         content_file_parser.cpp
-        effect_holder_groups_file_parser.cpp
-        string_groups_file_parser.cpp
 )
 
 target_sources(dndmanager_tests
     PRIVATE
-        content_parser.cpp
         content_file_parser.cpp
-        effect_holder_groups_file_parser.cpp
-        string_groups_file_parser.cpp
 )
 
+add_subdirectory(controllers)
 add_subdirectory(models)
