+++ src/parsing/controllers/CMakeLists.txt
@@ -1,0 +1,13 @@
+target_sources(dndmanager
+    PRIVATE
+        content_parser.cpp
+        effect_holder_groups_file_parser.cpp
+        string_groups_file_parser.cpp
+)
+
+target_sources(dndmanager_tests
+    PRIVATE
+        content_parser.cpp
+        effect_holder_groups_file_parser.cpp
+        string_groups_file_parser.cpp
+)
