+++ src/parsing/controllers/effect_holder_groups_file_parser.cpp
