+++ src/parsing/controllers/string_groups_file_parser.cpp
