+++ tests/parsing/controllers/content_parser_test.cpp
@@ -1,4 +1,4 @@
-#include "parsing/content_parser.hpp"
+#include "parsing/controllers/content_parser.hpp"
 
 #include <catch2/catch_test_macros.hpp>
